atualizando tratamentos e adicionando o dto

+++ src/main/java/com/residencia/Academia/controller/AtividadeController.java
@@ -14,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.residencia.Academia.dto.AtividadeDTO;
 import com.residencia.Academia.entity.Atividade;
+import com.residencia.Academia.exception.NoSuchElementFoundException;
 import com.residencia.Academia.service.AtividadeService;
 
 @RestController
@@ -29,31 +31,51 @@ public class AtividadeController {
 		return new ResponseEntity<>(atividadeList, HttpStatus.OK);
 	}
 
+	@GetMapping("/dto/{id}")
+	public ResponseEntity<AtividadeDTO> findAtividadeDTOById(@PathVariable Integer id) {
+		if (null == atividadeService.findAtividadeDTOById(id).getIdAtividade())
+			throw new NoSuchElementFoundException("Não foi encontrada Atividade com o id " + id);
+		else
+			return new ResponseEntity<>(atividadeService.findAtividadeDTOById(id), HttpStatus.OK);
+	}
+
 	@GetMapping("/{id}")
 	public ResponseEntity<Atividade> findAtividadeById(@PathVariable Integer id) {
-		// return new ResponseEntity<>(atividadeService.findAtividadeById(id),
-		// HttpStatus.OK);
 		Atividade atividade = atividadeService.findAtividadeById(id);
 		if (null == atividade)
-			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+			throw new NoSuchElementFoundException("Não foi encontrado Instrutor com o id " + id);
 		else
 			return new ResponseEntity<>(atividade, HttpStatus.OK);
 	}
 
 	@PostMapping
 	public ResponseEntity<Atividade> saveAtividade(@RequestBody Atividade atividade) {
-		Atividade novoAtividade = atividadeService.saveAtividade(atividade);
-		return new ResponseEntity<>(novoAtividade, HttpStatus.CREATED);
+		Atividade novaAtividade = atividadeService.saveAtividade(atividade);
+		return new ResponseEntity<>(novaAtividade, HttpStatus.CREATED);
+	}
+
+	@PostMapping("/dto")
+	public ResponseEntity<AtividadeDTO> saveAtividadeDTO(@RequestBody AtividadeDTO atividadeDTO) {
+		AtividadeDTO novaAtividadeDTO = atividadeService.saveAtividadeDTO(atividadeDTO);
+		return new ResponseEntity<>(novaAtividadeDTO, HttpStatus.CREATED);
 	}
 
 	@PutMapping
 	public ResponseEntity<Atividade> updateAtividade(@RequestBody Atividade atividade) {
-		Atividade novoAtividade = atividadeService.updateAtividade(atividade);
-		return new ResponseEntity<>(novoAtividade, HttpStatus.OK);
+		if (atividadeService.findAtividadeById(atividade.getIdAtividade()) == null) {
+			throw new NoSuchElementFoundException("Não foi possível atualizar, não foi encontrada a Atividade com o id "
+					+ atividade.getIdAtividade());
+		}
+		return new ResponseEntity<>(atividadeService.updateAtividade(atividade), HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteAtividade(@PathVariable Integer id) {
+		Atividade atividade = atividadeService.findAtividadeById(id);
+		if (null == atividade)
+			throw new NoSuchElementFoundException(
+					"Não foi possível encontrar a Atividade, pois não existe Atividade com o id " + id);
+
 		atividadeService.deleteAtividade(id);
 		return new ResponseEntity<>("", HttpStatus.OK);
 	}
